taste_mate/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
httpx==0.26.0
openai==1.12.0
python-dotenv==1.0.1
google-genai